Implement VCardTest with phpunit framework instead of deprecated simpletest framework
